code fmt

+++ foundry/test/uniswap-v2/solutions/UniswapV2Swap.test.sol
@@ -35,8 +35,8 @@ contract UniswapV2SwapTest is Test {
         path[1] = DAI;
         path[2] = MKR;
 
-        uint amountIn = 1e18;
-        uint amountOutMin = 1;
+        uint256 amountIn = 1e18;
+        uint256 amountOutMin = 1;
 
         vm.prank(user);
         // Input token amount and all subsequent output token amounts
@@ -63,8 +63,8 @@ contract UniswapV2SwapTest is Test {
         path[1] = DAI;
         path[2] = MKR;
 
-        uint amountOut = 0.1 * 1e18;
-        uint amountInMax = 1e18;
+        uint256 amountOut = 0.1 * 1e18;
+        uint256 amountInMax = 1e18;
 
         vm.prank(user);
         // Input token amount and all subsequent output token amounts
